switch from npm to bun
bun is the cool hot new thing of the moment and really looks promissing.
Wth its version 1.0 being released I wanted to try to see if it could
decrease the build time in the CI/CD pipeline.

Artifact uploading time can fluctuate a lot and is by far the slowest step, so
in the end bun brought no tangible benefits to our deploy time.
At the very least, it is not slower than npm and there is the added benefit of
making a few things faster locally. With all this in mind, I'm not reverting
back but I'm also not married to bun.

A better approach for speeding up ci/cd builds could be caching the
astro output (stuff like the processed images), since locally it takes
only a few seconds to build versus the minutes in the ci/cd pipeline.
However I'm not sure if it would fit the free tier space that github
offers. Besides that, it would at max save 1 minute from the up to 10
minutes it takes to upload the artifacts.

Lastly, I still use npm to install sharp because it doesnt seem to be
recognized by astro when installing it with bun. I think it is part of
bun's growing pains, since until recently it was not possible to install
sharp with bun.