popup photos for table
Aparrently I cant use the async function because we need an html string,
not a promise.
So im using the preprocessed smaller pictures of each tombstone as a
static sibgle photo for each popup.
Better than nothing since some of those pictures are present in the
table but not on the map. If i find another way in the future I may try
the carousel again but for now this seems fine