stop FOUC with small hack
I would sometimes get a flash of an unstyled page before the css
finished loading. This would be accompanied by the following firefox
error:
Layout was forced before the page was fully loaded.
If stylesheets are not yet loaded this may cause a flash of unstyled content.

After searching it on google, his may not be my fault but actually a
firefox bug. Regardless, there are many reports that the simple hack of
adding a dummy script at the beggining of the body tag solves the issue.
source: https://stackoverflow.com/a/57888310/12092477