resize Map and svg inside it
on smaller screens the map would occupy everything and there would be no
way to scroll back.

besides that, I resized the svg so that it would always be centered and
fully shown from the beggining.

lastly, on bigger screens I set the width to fit the svg size, so in
bigger screens the map border will always perfectly contour the map,
kinda like a square convex hull if thats a thing