add i18n fr, ptbr to tables
added translations for the tables in french and portuguese, however I
cant seem to find a way to easily translate the table columns.
Tabulators localization functions dont seem to work, be it through
local=true or by forcing a location with set locale.

Maybe transfer all the code to the astro cemetery main body and pass
the dictionary as a server variable to the script tag?