fix js not loading depending on navigation
view transition is not implemented on every browser yet, so it can get
buggy when navigating since it believes its all a really big page and
the js has been already executed.

got some help on the astro discord for this one, took a long time to
track down. Thanks arsh and sarah11918!

Here are 2 usefull references:
https://docs.astro.build/en/guides/view-transitions/#astropage-load
https://docs.astro.build/en/tutorials/add-view-transitions/#update-scripts

TODO: once view transition support is mature and more widely supported
this commit could be reverted