js: dinamically/async import js
the idea is that by creating an async import we essencially offload
some if the processing into after the main page is loaded
and decreasing the critical chain.

I also took the opportunnity to refactor some of the cemetery elements
into more componentish elements on their own folder and extracting most
js into js files.

Apparently another alternative could have been the use of partytown,
but since now this is done I see no reason to switch to it if this
solution works.

PS: an added bonus of moving the table logic into its own component
was the fix of an at first unnoticed bug where in translated pages
it wouldn't be able to find the js file to import since it would be
a level deeper.

Thanks wassfila and, once again, arsh for the tips.
Useful reference:
https://github.com/MicroWebStacks/astro-examples#09_dynamic-imports