fixbug: js not loading on first access
last commit aparently introduced an issue where js wasnt loaded
on first access to the cemetery page because astro:load
doesnt run (I think) if it is not a viewtransition navigation.

So I made verything run twice. Certainly not the cleanest solution, but
the fastest when accounting for research time. Since this will all be
uselles when view transitions gets proper browser support, I'm not very
concerned.

Adittionally, I moved the rest of the tooltip code into the js file and
encapsulated everything with a single function.