carouseel: add 'loading' background
sometimes you may be using a carousel but the next img hasnt been loaded
yet. By adding it as a background, we give the user feedback as to whats
going on without any performance impact and improving UX.

additionally, I made the tooltip loading state idemtical to the carousel
one in an attempt to make the transition more smooth and discreet
without resizing of elements.
To do that I had to fine tune the tooltip and carousel sizes, which was
awful and janky to do. The arbitrary looking widths and heights are to
take into account tippy paddings and other margins/paddings "out of our
control". Preliminary tests indicate we achieved consistent behaviour
across screensizes, so any refactoring to this section should be made
with extreme care since there is no obvious documentation to help us
re-align those loading states.