rm ViewTransitions
viewtransitions are not yet widely support and dealing with them in
firefox has been awful and buggy.

Removing it will decrease clutter and, according to our preliminary
tests, also fix the cemitery popup bug where clicking for the next
picture will trigger navigation but not update the image (only in the
first time).
This bug is hard to consistently replicate so debugging it was a
challenge, the biggest clue being that navigation appeared to affect it.