redraw the svg cemetery by hand
the old map was generated from a png of a CAD file (that to the best of
my knowledge was not precise, it was used more as an image editor than a
CAD software). The process of vetorizing the already imprecise png and
other further manual optimizations led to a file that looked patchworked
due to the font used not meshing well with the vectorized characters
that came from processing the png.
Besides that, converting the text characters to curves was a mistake,
increasing the number of nodes unecessarilly.

The new svg was created by tracing the old one in inkscape and recreating
it from scratch. It isnt a "pixel perfect" recreation, but the original
also wasnt a precise representation of reality, serving more to give a
general idea of the lay of the land. I also altered a bit the geometry
of some tombs to better fit their ID, unlike the original we
consistently utilized the same font size for the entirety of the svg.

This new iteration also colored red all tombs without photos available.